fix: make sure to create picker before calling update, fixes #274
