+++ demo/pages/input/index.html
@@ -12,6 +12,7 @@
 	<body class="font-sans bg-white bg-light-mode text-slate-900 min-h-screen container mx-auto text-center flex flex-col items-center py-12 dark:bg-slate-900 dark:bg-dark-mode dark:text-white">
 		<h1 class="block mb-7 text-6xl font-extrabold">Vanilla Calendar</h1>
 		<p class="text-lg mb-12 dark:text-slate-400">A pure JavaScript date and time picker using TypeScript so it supports any JS framework and library.</p>
+		<button id="set-date" type="button" class="mb-7 w-64 border-gray-200 border-2">Set Date</button>
 		<label>
 			<div>Datepicker tag 'input'</div>
 			<input id="calendar-input" class="input" name="calendar" type="text" readonly />
